+++ pkgs/native_assets_builder/test_data/wrong_namespace_asset/hook/build.dart
@@ -16,7 +16,7 @@ void main(List<String> arguments) async {
     }
 
     output.addAsset(
-      NativeCodeAsset(
+      CodeAsset(
         package: 'other_package',
         name: 'foo',
         file: assetUri,