+++ pkgs/objective_c/lib/objective_c.dart
@@ -90,3 +90,4 @@ export 'src/objective_c_bindings_generated.dart'
         NSValue,
         Protocol;
 export 'src/protocol_builder.dart';
+export 'src/selector.dart';