+++ pkgs/objective_c/CHANGELOG.md
@@ -12,6 +12,7 @@
   bindings if an optional method is invoked, and the instance doesn't implement
   the method.
 - Dispatch all object/block releases to the main thread.
+- Add utils for converting Dart `String`s to Objective-C selectors and back.
 - Require Dart 3.4 or later (due to the use of `dart:ffi`
   `Struct.create` by `package:ffigen`).
 - __Breaking change__: Return structs from ObjC methods by value instead of